requirements.txt
django==3.1.1
pylint-django
djangorestframework==3.11.1
drf-yasg
markdown
psycopg2-binary
pydotenv
pygit2
pygithub
pygount
scikit-learn
unidecode
perceval
